Unambiguously Bad Include Hack.
Loading a template without parsing it has no immediately clear solution.
The {% ssi %} tag permits this, but requires the use of absolute include
paths, which destroys the portability of this application.

Becauuse the default Template loaders parse the template and return a
Template object instead of the raw text of the file, options are limited:
#1 Write a new template loader method to load a template without parsing it
#2 Write a render method that re-combines parsed Template nodes back into
raw text or a VerbatimNode
#3 Write a quick and dirty template tag to load any arbitrary template from
the filesystem.

This is option #3.

Signed-off-by: David Arbuckle <dintlu@gmail.com>