For loops working for arrays, arrays of arrays, and dicts.
Because JS loops only support declaration of the key within the looping
syntax, additional variable declarations are made within the body of the
loop itself.

I'm not sure whether Django executes for loops in a separate
namespace in order to prevent inner declarations from overwriting outer
variables, but that isn't done in this case.

Signed-off-by: David Arbuckle <dintlu@gmail.com>