Implemented {% empty %} tag.
Because the template parser doesn't look ahead for the closing node, we
have to inspect the loop before it is entered, and account for the possible
empty loop in javascript after the fact.

The loop object or array is checked for length, and an IF condition
permits the for loop to execute.  When the IF condition evaluates false,
the empty tag will populate an ELSE tag for the condition.  In order to
ensure that the correct number of closing } are present for the logic,
the array mango.tags._in_empty will have a value pushed to it so that the
parser is aware.

This solution is a little bit convoluted, but I believe it is preferable to
 the alternative of saving the entire loop string until an {%empty%}
or {%endfor%} tag is reached.

Signed-off-by: David Arbuckle <dintlu@gmail.com>