Implemented a small variety of filters.
These filters still need to be tested to ensure that they are reliably
chainable, and that the behavior when used with various data types is
consistent with Django's current handling.

Some current problems:
linebreaksbr - the <br> tags need to be rendered as HTML, while the rest
fof the var should be rendered as text.
upper, lower - when used on a list, these are not mapped to all the list
values.

Signed-off-by: David Arbuckle <dintlu@gmail.com>