Re-implemented Upper and Lower so that the transformations would be applied over entire objects without converting those objects to strings first.
The implementation is still flawed with respect to the Django version, in
that dictionary keys are not converted to uppercase.  In addition, a helper
function was created to traverse objects and apply string transformations,
but I believe that it may be a use-case-specific implementation of
Array.map(), and may benefit from a refactoring.

Signed-off-by: David Arbuckle <dintlu@gmail.com>