Cycle tag is now context sensitive; Applies filters; Applies silent;
Supports every weird variation of cycle that Django templates do with 1:1
feature parity.

Moved nearly all of the code related to instantiation of the local _cycle
object into the template; that is, when the template is compiled it will
instantiate new mango.cycle() objects for each of the cycles in the template.

Whenever a template begins rendering, the obj._cycles dictionary is reset
in order to prevent previous cycles from leaking into the current namespace.

This is the most complex feature implemented so far.
The worst of the insanity is that it continues to use eval() statements in
order to render the filtered variables, and that a regexp is used to parse
the cycle tag's arguments in order to handle the additional complexity
added by filter support.

That said, I may attempt to remove the reliance on eval(), but I am
skeptical whether it is a worthwhile endeavor at this time.

Signed-off-by: David Arbuckle <dintlu@gmail.com>