Implemented {% filter %}.
When {% filter %} is encountered, a closure is created in template function
into which the filter values are passed as arguments.  All of the text that
is rendered within this closure is then passed to mango.filter.apply() with
evaluate_filters = true;

mango.filter.apply() has been updated to evaluate all filters against the
passed string when evaluate_filters=true. When this argument is passed,
the tvar is *not* split to search for filters, instead the tag_filters
argument is checked for comma-separated filters, which are applied to the
raw tvar.

There is currently a defect with nested filters, where the outermost filter
action does not override inner filters.




Signed-off-by: David Arbuckle <dintlu@gmail.com>