Implemented mango.is_true method to evalueate Truthiness in a Pythonic way.
In Javascript, empty lists evaluate to "true": [] == true
In Python, empty lists evaluate to "False":  [] == False

In Javascript, empty dicts evaluate to "true": {} == true
In Python, empty dicts evaluate to "False": {} == False

In Javascript, True is undefined.  We get around this by declaring local
variables for True and False in the rendering function.  This was added
outside of the scope of this commit.

Signed-off-by: David Arbuckle <dintlu@gmail.com>