Added date filter.
In the process, encountered error with Django's own parsing of "W"-format
dates, in addition to issues converting naive datetime strings to new Date()
objects in Javascript.

When a new date is created in JS, the timezone is implicitly defined as UTC
and the date is cast to display in the client's local time.  This will need
to be resolved ASAP.

Signed-off-by: David Arbuckle <dintlu@gmail.com>