Initial support for forloop inner vars.
This implementation lacks support for the parentloop syntax, and nesting
loops will result in the inner loop overwriting the value of the outer
loop.  The 'simple' solution of copying the parent forloop object to
parentloop doesn't work, since that only copies a reference.

More thought will need to be given to surmounting this obstacle. I am
leaning towards a prototype-based approach where new forloop objects are
stored in an array, and are pushed or popped  whenever a loop is opened or
closed.

Signed-off-by: David Arbuckle <dintlu@gmail.com>